changed my code structure and edited my submitform
