--- src/components/PostThoughts.css
@@ -1,7 +1,0 @@
-.input-wrapper {
-	border: 1px black solid;
-	display: flex;
-	flex-direction: column;
-	font-size: 30px;
-	padding: 40px;
-}
